requirements.txt
Flask
pylint
Pandas
Numpy
lxml
discord.py
yfinance
aiocron
plotly
matplotlib
python-dotenv
aiohttp
aiosignal
astroid
async-timeout
attrs
blinker
certifi
chardet
charset-normalizer
click
dill
frozenlist
idna
isort
itsdangerous
Jinja2
MarkupSafe
mccabe
multidict
platformdirs
requests
slack-sdk
tomli
tomlkit
typing-extensions
urllib3
Werkzeug
yarl
flask-marshmallow
Flask-SQLAlchemy
marshmallow
SQLAlchemy
